+++ jarvis-web/src/routes/house/dashboards/+page.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import Dashboards from '$lib/sections/Dashboards.svelte';
-</script>
-
-<Dashboards />